+++ packages/database/drizzle/schema/postgres.ts
@@ -30,8 +30,6 @@ export const user = pgTable("user", {
 		.defaultNow()
 		.$onUpdate(() => /* @__PURE__ */ new Date())
 		.notNull(),
-	username: text("username").unique(),
-	displayUsername: text("displayUsername"),
 	role: text("role"),
 	banned: boolean("banned").default(false),
 	banReason: text("banReason"),